Test: (BRD-126) 이메일 유효성 확인 API 명세 문서화(RestDocs)
